Enable UnpackStrategy types
