+++ Library/Homebrew/extend/object.rbi
@@ -1,0 +1,6 @@
+# typed: strict
+
+class Object
+  sig { returns(T::Boolean) }
+  def present?; end
+end
