+++ Library/Homebrew/unpack_strategy/air.rb
@@ -8,8 +8,6 @@ module UnpackStrategy
 
     include UnpackStrategy
 
-    using Magic
-
     sig { returns(T::Array[String]) }
     def self.extensions
       [".air"]
