+++ Library/Homebrew/unpack_strategy/bazaar.rb
@@ -8,7 +8,7 @@ module UnpackStrategy
   class Bazaar < Directory
     extend T::Sig
 
-    using Magic
+    
 
     def self.can_extract?(path)
       super && (path/".bzr").directory?
