+++ Library/Homebrew/unpack_strategy/compress.rb
@@ -8,7 +8,7 @@ module UnpackStrategy
   class Compress < Tar
     extend T::Sig
 
-    using Magic
+    
 
     sig { returns(T::Array[String]) }
     def self.extensions
