+++ Library/Homebrew/unpack_strategy/cvs.rb
@@ -6,7 +6,7 @@ require_relative "directory"
 module UnpackStrategy
   # Strategy for unpacking CVS repositories.
   class Cvs < Directory
-    using Magic
+    
 
     def self.can_extract?(path)
       super && (path/"CVS").directory?
