+++ Library/Homebrew/unpack_strategy/fossil.rb
@@ -11,7 +11,7 @@ module UnpackStrategy
     include UnpackStrategy
     extend SystemCommand::Mixin
 
-    using Magic
+    
 
     sig { returns(T::Array[String]) }
     def self.extensions
