+++ Library/Homebrew/unpack_strategy/git.rb
@@ -6,7 +6,7 @@ require_relative "directory"
 module UnpackStrategy
   # Strategy for unpacking Git repositories.
   class Git < Directory
-    using Magic
+    
 
     def self.can_extract?(path)
       super && (path/".git").directory?
