+++ Library/Homebrew/unpack_strategy/lua_rock.rb
@@ -8,7 +8,7 @@ module UnpackStrategy
   class LuaRock < Uncompressed
     extend T::Sig
 
-    using Magic
+    
 
     sig { returns(T::Array[String]) }
     def self.extensions
