+++ Library/Homebrew/unpack_strategy/mercurial.rb
@@ -6,7 +6,7 @@ require_relative "directory"
 module UnpackStrategy
   # Strategy for unpacking Mercurial repositories.
   class Mercurial < Directory
-    using Magic
+    
 
     def self.can_extract?(path)
       super && (path/".hg").directory?
