+++ Library/Homebrew/unpack_strategy/microsoft_office_xml.rb
@@ -8,7 +8,7 @@ module UnpackStrategy
   class MicrosoftOfficeXml < Uncompressed
     extend T::Sig
 
-    using Magic
+    
 
     sig { returns(T::Array[String]) }
     def self.extensions
