+++ Library/Homebrew/unpack_strategy/otf.rb
@@ -8,7 +8,7 @@ module UnpackStrategy
   class Otf < Uncompressed
     extend T::Sig
 
-    using Magic
+    
 
     sig { returns(T::Array[String]) }
     def self.extensions
