+++ Library/Homebrew/unpack_strategy/pkg.rb
@@ -8,7 +8,7 @@ module UnpackStrategy
   class Pkg < Uncompressed
     extend T::Sig
 
-    using Magic
+    
 
     sig { returns(T::Array[String]) }
     def self.extensions
