+++ Library/Homebrew/unpack_strategy/self_extracting_executable.rb
@@ -8,7 +8,7 @@ module UnpackStrategy
   class SelfExtractingExecutable < GenericUnar
     extend T::Sig
 
-    using Magic
+    
 
     sig { returns(T::Array[String]) }
     def self.extensions
