+++ Library/Homebrew/unpack_strategy/sit.rb
@@ -8,7 +8,7 @@ module UnpackStrategy
   class Sit < GenericUnar
     extend T::Sig
 
-    using Magic
+    
 
     sig { returns(T::Array[String]) }
     def self.extensions
