+++ Library/Homebrew/unpack_strategy/subversion.rb
@@ -6,7 +6,7 @@ require_relative "directory"
 module UnpackStrategy
   # Strategy for unpacking Subversion repositories.
   class Subversion < Directory
-    using Magic
+    
 
     def self.can_extract?(path)
       super && (path/".svn").directory?
