+++ Library/Homebrew/unpack_strategy/ttf.rb
@@ -8,7 +8,7 @@ module UnpackStrategy
   class Ttf < Uncompressed
     extend T::Sig
 
-    using Magic
+    
 
     sig { returns(T::Array[String]) }
     def self.extensions
